New library SDKs need to be pushed to github.
Note that the sources.jar is pushed. Coursier caching does not yet automatically downloading. Thus hyper-cms links to the shadow sources.jar in the local custom repository that shadows the custom repository. This is ugly and incorrect but seems to be a bug in the coursier cache SBT plugin, about which I won't worry right now.

There was just an update to the SBT plugin a little while ago. Perhaps the problem has been fixed.