I modified the installation scripts to change the layout of the repositories and the cache.
This layout seems to be correct even if it is not really document. It is partially documented because one Apache reference states that the <xml> of the pom can be used on the command line.

We are stuck with the command line because we are using build.sbt and not pom.xml.